Final commit

+++ Sellers Application/app/src/main/java/ru/eg/sellersapplication/data/pojo/Amount.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Amount(
     @SerializedName("currency") val currency: String,
-    @SerializedName("value") val value: Int
+    @SerializedName("value") val value: Double
 )
